Permissions of Accessing Other VM’s Memory change to "permissions in accessing other VMs' memory"
Since entire (add the) target VM is not trusted and no changes should be made to the target VM

to remap memory pages from (add the) target VM to its own memory space.

by (add the) Xen Security Module

For example, the permission of map_read allow (change to allows) a domain to map other domain’s memory with read only (change to read-only) permission.

In addition to the permissions granted from XSM, we also need to (delete to) the permission (add to here and delete the) the read information from Xenstore

The next challenge is building new libraries into Mini-OS. Mini-OS is a examplar (change to an exemplary?) minimal operating system designed to running (change to run) on Xen hypervisor.

To keep the kernel small, there are only (add a) few libraries (add that) can be built in it:

To built (change to build) those code (either these codes or this code) in Mini-OS, we need to cross compile C++ standard libraries into the tiny kernel.

with permissions to introspect other (change other to another) guest VM via VMI technique.